Configure Webpack to create release and tag it in the Frontend
Along with enabling integrations in github and vercel, this change
allows for commits to be associated with deploys in sentry.io.

In local environments, including `vercel dev`, the webpack plugin
defaults to "dry run" mode and doesn't attempt to create a release or
associate commits in Sentry. Commits are only tied to a release in
the specified config when vercel deploys from the "production" branch
in this configuration.

test plan:

The "production" branch head in https://github.com/topocount/coordinape
is currently throwing an error on pageload. That error is associated
with the tagged release (represented by the commit SHA) which is
associated with recent commits.