Create Hasura createUser Action
I've defaulted to camelCase naming within graphQL because
that's the convention, but I'm happy to go back to snake case if that's
preferred.

I've grouped all potential updates/inserts into one mutation,
to minimize the i/o and better utilize graphQL.

I still need to write the authentication code. Because the serverless
function is hosted publicly by vercel, we should auth on a pre-shared
environmental secret that'd be passed as a request header. We should
probably be doing this on all vercel-hosted actions to ensure we don't
see unauthorized db writes. Only hasura should be able to do anything
with these functions.