fix spoofing address issue for create circle (#526)
* Create Hasura createUser Action

I've defaulted to camelCase naming within graphQL because
that's the convention, but I'm happy to go back to snake case if that's
preferred.

I've grouped all potential updates/inserts into one mutation,
to minimize the i/o and better utilize graphQL.

I still need to write the authentication code. Because the serverless
function is hosted publicly by vercel, we should auth on a pre-shared
environmental secret that'd be passed as a request header. We should
probably be doing this on all vercel-hosted actions to ensure we don't
see unauthorized db writes. Only hasura should be able to do anything
with these functions.

* Add validation that the request originated from hasura

* Refactor zEthAddress so it plays nicely with safeParseAsync

the ethersjs `provider.resolveName` function throws on a malformed
eth address so it needs to be refined before it is passed in.

As a bonus, failed ENS resolutions and malformed eth address inputs
now throw different errors

* Incorporate Review feedback

- move env var to config.ts
- change address lookups from `_eq` to `_ilike`

* Further review-related fixes

- use sound form of if statement
- return all params that were requested
- validate that the requesting user is a circle admin

* Create Composable response body zod parser

* Make parsing "strict" instead of "strip" so an error is thrown

* fix spoofing address issue for create circle

- remove address as an input

* remove default value for coordinape user address

* add middleware and zod

* remove lc

* run linter

* HasuraRequestBody composer works with trailing refines and transforms

* yarn generate

Co-authored-by: Kevin Siegler <17910833+topocount@users.noreply.github.com>