change if-condition to assert
the function will never be called when the if-condition is false, since form validation will fail beforehand. so we can use assert to 1) ensure this is the case at runtime & 2) communicate this expectation to readers of the code

also, in this commit, I removed the fallback to 18 decimal places. this is already being done in useVaultFactory