Update profile/skills tags (#653)
* Update profile/skills tags

This adds a few tags that we want to add to the app for Yearn's sake. We worked to make them generic enough to be useful to others and not be weird or out of place. Resolves #644

* Added requested asterisks.

whoops, forgot asterisks

* put new skills first in list, remove asterisks

If these will be first in the list on view, then that will make them very easy to find. I checked with Zemm and he agrees that this method should make them easy for the Yearn team to work with, but also these will blend in more nicely with our "normal" skills tags.