bugfix: remove dom dependency for NodeJS ts code (#745)
We were intiially relying on the frontend tsconfig.json file to
typecheck our api handlers. This is bad because we're expecting dom
libraries to be accessible in our backend ts code. This led to some
funky things like "shimming" node-fetch into our code when it should
work out of the box.

test plan:

1. vercel dev
2. ensure webpack is able to build
3. test some endpoints but this isn't technically necessary
4. `yarn db-seed-fresh` just to be safe.