Create a Contributing Doc
I added a lot of low-key opinionated ideas and concepts, but they're all
for the purpose of making the review process simpler, and believe these
are all practices we should be working towards.

It's worth noting that much of this is borrowed from SourceCred's
CONTRIBUTING doc, with the language softened a bit.