fix: add explicit timestamptz type to actions.graphql
timestamptz is a builting scalar type in hasura. However, it is not a
builtin in graphql, so our linter fails to parse the file. Explicitly
declaring it allows for the file to be parsed, linted, and fixed again.

Unfortunately, it's not capital-cased type, so I've had to add a rule to
the .eslintrc config file to prevent it from throwing a lint error.

test plan:
- yarn lint:check
- yarn hasura metadata apply
- yarn hasura metadata export
- ensure gql queries using the timestamptz type continue to behave as
  expected.