Opt In Quality of Life Improvements
These changes make it easier for users to opt in and for circle creators
to onboard a circle faster.

First, created users used to be opted out upon creation, and had to opt
in during their first epoch to begin receiving GIVE. Now they are
opted in by default in our app, and there is no longer this chicken and
egg issue of users opting in and having no one to GIVE to because others
are not opted in.

Second, the UI now has the Opt-in radio button selected on every page
load. This makes it easier to opt in, and the user is now required to
explicitly opt out before receiving give.

Test plan:

E2E tests should still pass and unit tests still pass

Related Issue:
1161