Add Pending Vault Txs Table
This table is utilized to track pending transactions for app
interactions that require backend updates once mined. This class of
transactions includes vault creation, distribution creation, and claims.

This is part of a larger feature containing a cron task that will check
these pending transactions periodically and update the backend state
appropriately once mined. Once a transaction completes and is handled
the pending tx will be deleted, either by the frontend or the backend.