fix errors about passing refs to function components
all this commit does is wrap some components in React.forwardRef

i refactored the props-handling just to minimize the diff. otherwise, when you wrap components in React.forwardRef, it indents the whole component

couldn't avoid that in DeprecatedApeTextField though; had to use a generator function instead of an ordinary argument because `ref` is specially handled

couldn't just wrap all of DeprecatedApeTextField in React.forwardRef either, because the material-ui components in FormTimePicker and FormDatePicker need it to be unwrapped. so i created a new wrapped version, DeprecatedApeTextFieldWithRef, which is an abomination of a name, but hey, this is deprecated code anyway