Fix FormTokenField Decimal Handling
FormTokenField was handling token decimals incorrectly. This fix makes
it impossible to enter a value with more decimals than a token has
available. The amount will remain at the value with the smallest
possible decimal. If a user pastes in a value with more decimals than is
can be handled, the amount is truncated to the valid decimal quantity.

Additionally, this refactor removes that annoying thing where the field
can't be empty and you need to "type around" a placeholder zero.
Validation treats empty fields correctly in my testing on deposits,
           withdraws and distributions.

This change also led to an upstream refactor to move all calculations
away from Numbers and to BigNumbers in all views that consume
FormTokenField.

Test Plan: unit and E2E tests should pass.