Add Tests
The controller seems to ignore input html configs and allows any values
in the box, so we need to manually remove invalid characters? I haven't
seen this type of behavior before but a solution was easy enough with
regex.

It's probably because we were using strings in a number-type input