Fix FormTokenField Decimal Handling (#1259)
* Fix FormTokenField Decimal Handling

FormTokenField was handling token decimals incorrectly. This fix makes
it impossible to enter a value with more decimals than a token has
available. The amount will remain at the value with the smallest
possible decimal. If a user pastes in a value with more decimals than is
can be handled, the amount is truncated to the valid decimal quantity.

Additionally, this refactor removes that annoying thing where the field
can't be empty and you need to "type around" a placeholder zero.
Validation treats empty fields correctly in my testing on deposits,
           withdraws and distributions.

This change also led to an upstream refactor to move all calculations
away from Numbers and to BigNumbers in all views that consume
FormTokenField.

Test Plan: unit and E2E tests should pass.

* Add Tests

The controller seems to ignore input html configs and allows any values
in the box, so we need to manually remove invalid characters? I haven't
seen this type of behavior before but a solution was easy enough with
regex.

It's probably because we were using strings in a number-type input

* Fix E2E tests

* Update Mock State to better reflect production schema

add required amount values