Create homebrewed dev stack server
Motivation

vercel was a great drop-in batteries-included dev server for our
serverless functions, but it required a bunch of external config that
was really painful to utilize without an internet connection, especially
when setting up a repo locally.

Details

This new dev server handles all requests to serverless functions, except
for cron tasks, and proxies all other requests to the webpack dev server.

test plan: This is a drop-in replacement for the existing `vercel dev`
command.

Future Work

Route(s) for the cron tasks need to be created at some point as work
evolves on those