Add some small waits
cypress is directed to click on labels during these tests, which in
normal environments, seemingly instantly transfer focus to its
associated `input` field. This delay should allow plenty of time for
that transfer of focus to occur.

The intention behind clicking the label was originally some notion of
accessibility, but if this is flaky enough (I've seen this error a
    handful of times over the months) it can be changed.