Refactor useContributions to be a bit more organized
Delete the deprecated one and use the new one in its place

This lays some groundwork for completely separating an integration's
logic from the process of invoking the integration and handling its
output in our app context. This makes the building of an "integrations
repo" for contributions fairly straightforward, codewise and allows us
to specify interfaces, should we want to for this implementation.

Test plan:

I tested this locally on the original allocations page using dework and
observed behavior in parity before deleting the deprecated
implementation, where I had a completed task show up in the UI that
linked back to the task in Dework.