Nav Button BugFixes
@sshmm discovered some buggy behavior with the current nav buttons when
jumping between contributions that are grouped into different epochs.

I fixed this by formalizing the "Latest" timespan as a pseudo-epoch of
sorts, and add it to the epochs array if we're outside of an epoch and
no future epoch is scheduled. The time logic was getting tricky so i
added a bunch of library functions that probably need unit tests at
some point to verify correctness of behavior.

test plan: Play around with seeded data to ensure that the nav buttons
behave correctly.

merge plan: rebase and merge after #1494