Hide irrelevant epochs and integrated contributions
Integrator APIs can return errors if our filters are too large. Intead
of querying for all contributions since the beginning of time for an
initial epoch in a circle, we now query only back to the month before
the epoch started.

In the Contributions Page, Epochs that ended before the Contributions
feature was launched are now hidden.

Potential Future work: optionally hide all epochs with no contributions