bugfix: Contributions doesn't re-render when data changes
There was an edge case when if you switch between one selected Contribution to another selected Contribution BEFORE the mutation finished saving, the useMemo would not be invoked, and the page would show old state even though the save succeeded. We change from that pattern to use a dataUpdatedAt memo pattern that properly updates when new data is fetched.