bugfix: contribution entry with many future epochs
Currently, if an epoch is scheduled beyond the current one,
contributions cannot be entered, since the frontend recognizes that
the latest epoch is not "current" and incorrectly assumes that it and all
subsequent epochs are in the past. These changes fix that logic.

Additionally, the label logic was misbehaving in a similar manner and
has been fixed as well.

test plan:

seed the db and create some future epochs in different circles.
Ensure contributions can still be entered and that the epoch dates
and labels are correct in the create contribution panel and on the
page as well.