bugfix: stop querying contributions from time 0
Contribution providers do not want to support multi-year queries due to
adversarial concerns such as scraping. Our contribution hook would fall
back to epoch timestamp zero if no start date for the query was
supplied. This is now fixed and the query falls back to a month before
the current epochs start date if no prior epoch exists.

I've also tightened up the API to make the dates required params so that
no hidden fallbacks are possible. This is at the expense of the legacy
code in the ProfileCard, which utilizes recoil and irresponsibly tries
to render a result even if no epoch is currently active.

It'd be preferable to pass in a required epoch prop or two that the
component can rely on, but given that this is a legacy view, it's not
worth the maintenance. The solution, then is nominally unreachable hack
that would yield a minimally impactful query for integrated
contributions