bugfix: don't perform React state updates on unmounted component
setSelectedChain is called in an async function, which sometimes runs after the component is no longer available, which causes an other similar React state update of an unmounted component error. We fix this issue by passing in a reference to if the component is mounted or not, and only invoking the state change if the component is still mounted.