Add Logger to the gql codegen
This allows us to more intricately control how zeus is logging out
console errors.

I've also used this as an opportunity to setup a robust example for
mocking out the logger in tests, and set up the jest infra that makes it
super easy to introspect on the mocked Logger behavior.

Tests written by @zashton have been updated accordingly