Implement New CreateEpoch API
This new API includes functionality for custom and one-off epochs to be
scheduled. monthly epochs (epochs that start and end on the same weekday
each month) will be implemented after the intial integration for
one-offs and custom repeating epochs is complete and they are beginning
to be tested

Test plan:
unit tests are provided

Merge plan:
This work will be merged into staging after the frontend integration is
complete, and QA testing can be done.