Refactor Time and Date Pickers to raw elements
This PR migrates away from MUI for these pickers.

However, there is a decision to be made about how we want the state to
work fro these components.

The MUI components accepted and statefully held ISO6801 DateTimes. This
meant that we could pass the same field into both of these components
and it would "just" work. However, since we are relying on the more
minimalist react-hook-form library to manage form state, reimplementing
this state is potentially more engineering work than we necessarily
need.

The alternative to get these components merged back down to main is to
just use a separate form field entry for the time and and then
reconstitute the ISO 8601 datetime as part of the upstream _form state_
as opposed the aforementioned _field state_.