feat: Create table to store api tokens for the discord bot and circle linking flow (#1788)
* Create table to store api tokens for the discord bot

We want to ensure the links generated to complete the flow are
single-use and do not cause unintended side-effects if the link is
accessed more than once. This flow guarantees that the links will be
single-use.

The schema and permissions support the following flow:

1. The discord bot receives a request to administer a circle via the
   discord bot. This might entail the following:
   - adding or removing users via discord roles
   - assigning the circle to a channel
   - configure circle notifications in discord
2. the bot inserts a row in the discord.circle_api_tokens table
   containing a circle_id and begins listening for the token to be added
   to that row via a subscription
3. the bot generates a link that will open the coordinape app on the
   discord route with the param `id` representing the row id in the
   circle_api_tokens table.
4. The user opens the link in a browser and authenticates or is
   authenticated with the address that has admin perms on the circle.
5. In the frontend, a circle api token is generated and passed into an
   action handler along with the circle_id provided, and the id of the
   token.
6. The handler attempts to update the row created by the bot with the
   new id. If no row exists with a null token field and the
   update fails, the key is deleted and an error is shown to the user
   in a toast on redirect.
7. If the update is successful, the user is notified the circle is
   linked to the channel in both the app in a toast and in discord
   because the subscription returns affirmatively.

This schema and prescribed flow would not be possible for updating the
api token, so it's best to request all permissions the bot needs up
front, rather than having the permissions granted to the bot evolve
without compromising on the single-use aspect.

We could consider generating and passing around a single-use identifier
but the flow gets a bit more complex. This could also be added later
without much concern for additional lift. YNGNI for now seems prudent
though

* Rename roles to server_roles in discord schema

* Regenerate ts bindings and gql schema

* feat: LinkDiscordCircle handler

Co-authored-by: Kevin Siegler <17910833+topocount@users.noreply.github.com>