Implement Epoch Monthly backend
This logic accepts a start date, end date, and week number. The start
date and end date must be the same weekday. They also must fall on the
week number of the month. The first week of the month is week 0.

The week number is a necessary extra parameter because it can't
necessarily be inferred from the date. If a following date overflows
into the following month, the inferred week number would cause all
following epochs to start at the beginning of the following month than
the end of the current month, as was originally intended