Refactor updateEpoch to utilize repeat_data
This refactoring utilizes the more configurable epoch structures.

Note: modifying a repeating epoch that has already begun is quite
difficult, so we have to convert that epoch to a one-off, then quietly
create a new repeating epoch using the original config after the
one-off. This flow is already fleshed out in the frontend, but it needs
to be re-integrated.

Test plan

Unit tests have been written. They are fully integrated