Implement Epoch Monthly backend (#1829)
This logic accepts a start date, end date, and week number. The start
date and end date must be the same weekday. They also must fall on the
week number of the month. The first week of the month is week 0.

The week number is a necessary extra parameter because it can't
necessarily be inferred from the date. If a following date overflows
into the following month, the inferred week number would cause all
following epochs to start at the beginning of the following month than
the end of the current month, as was originally intended. @CryptoGraffe
has suggested that we truncate the n > 4th occurrence in a month
to the last occurrence in a month to better align with how GCal's
monthly occurrence config works.

Unit tests have been updated accordingly.


---------

Co-authored-by: Lawrence Wang <l@lw.io>