Implement new backend cron logic for repeat_data schema
This PR contains the epoch cron that will create repeating epochs using
the new `repeat_data` schema. A timzone has been added to the schema to
minimize the complexity of the calculations on the backend. However, the
input of this timzone is pretty unopinionated. If an unsupported zone is
entered, or just a nonsense string is, the timezone is configured to be
UTC, which eliminates any shifts due to DST. Otherwise, DSTs will be
observed when the timezone is applied to a datetime.

Merge Plan

This implementation is meant to be merged before the migration and
cutover, since the code supporting the old epoch logic is still present
and the encasing cron logic will function exactly the same until
the `repeat_data` column is populated. At that point, the new repeating
logic will kick in.