Bugfix: set timezone before calculating date in tests (#1941)
* Bugfix: set timezone before calculating date in tests

Daylight savings needs to be accounted for in start and end date
calculation when a time zone with DST is used. This was an issue with
the tests, not the actualy code

* Bugfix: Epcoh repeat migration (#1940)

The repeat data was missing the `type` member

I also improved the error message so we an know exactly what went wrong
without investigating the database.