Implement Variable Dust Allowance
Because the evm deals in fixed-point arithmetic, the order of magnitude
of the dust allowance should increase with the quantity of decimals a
token has.

For example, it does not make sense for DAI (18 decimals) to have the
same dust allowance as USDC (6 decimals) since that would mean we are
allowing 12 orders of magnitude less dust for DAI, making it much harder
to submit a distribution.

Test Plan

Unit Tests are  included and existing tests are updated.