Allow Api Tokens to access User Admin actions (#1945)
* Allow Api Tokens to access User Admin actions

Our auth logic allows circle api keys access to these actions, but these
explicit hasura-level permissions are missing.

* enable API access for createUsers, deleteUser, deleteUsers

This is all guarded by the newly created `manage_users` permission.

test plan

tested in graphIQL with a valid and invalid token on all endpoints. A
token without the explicit perm and a token for a different circle
behaved as expected.

* readd guild info perm